Split matrix_ops_test into static and dynamic tests
